# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/schedules": {
      "get": {
        "description": "Returns all schedules, optionally filtered by scheduleIds.",
        "operationId": "ScheduleController_findAll",
        "parameters": [
          {
            "name": "scheduleIds",
            "required": false,
            "in": "query",
            "description": "Comma-separated list of schedule IDs to filter by",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ScheduleResponseDto"
                  }
                }
              }
            }
          }
        },
        "summary": "Get all schedules",
        "tags": [
          "Schedule"
        ]
      },
      "post": {
        "operationId": "ScheduleController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateScheduleDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Schedule"
        ]
      }
    },
...
    "/api/schedules/{id}/registrations": {
      "post": {
        "operationId": "RegistrationController_create",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateRegistrationDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Registration"
        ]
      },
      "get": {
        "operationId": "RegistrationController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Registration"
        ]
      }
    },
    "/api/schedules/{id}/registrations/{studentId}": {
      "delete": {
        "operationId": "RegistrationController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "studentId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "tags": [
          "Registration"
        ]
      }
    },
    "/api/payments": {
      "get": {
        "operationId": "PaymentController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PaymentResponseDto"
                  }
                }
              }
            }
          }
        },
        "summary": "Get all payments",
        "tags": [
          "payments"
        ]
      },
      "post": {
        "operationId": "PaymentController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePaymentDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaymentResponseDto"
                }
              }
            }
          }
        },
        "summary": "Create a payment",
        "tags": [
          "payments"
        ]
      }
    },
...
    "/api/payments/{id}": {
      "get": {
        "operationId": "PaymentController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaymentResponseDto"
                }
              }
            }
          }
        },
        "summary": "Get a payment by ID",
        "tags": [
          "payments"
        ]
      },
      "patch": {
        "operationId": "PaymentController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePaymentDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaymentResponseDto"
                }
              }
            }
          }
        },
        "summary": "Update a payment",
        "tags": [
          "payments"
        ]
      }
    },
    "/api/transactions/me/credit-balance": {
      "get": {
        "operationId": "TransactionsController_getMyCreditBalance",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreditBalanceResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Transactions"
        ]
      }
    },
    "/api/transactions/{userId}/credit-balance": {
      "get": {
        "operationId": "TransactionsController_getCreditBalance",
        "parameters": [
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreditBalanceResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Transactions"
        ]
      }
    },
    "/api/transactions": {
      "get": {
        "operationId": "TransactionsController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TransactionResponseDto"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Transactions"
        ]
      },
      "post": {
        "operationId": "TransactionsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTransactionDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TransactionResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Transactions"
        ]
      }
    },
    "/api/transactions/me": {
      "get": {
        "operationId": "TransactionsController_findMy",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TransactionResponseDto"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Transactions"
        ]
      }
    },
    "/api/transactions/{userId}": {
      "get": {
        "operationId": "TransactionsController_findByUserId",
        "parameters": [
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TransactionResponseDto"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Transactions"
        ]
      }
    },
    "/api/products": {
      "get": {
        "operationId": "ProductController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": "List of products",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ProductResponseDto"
                  }
                }
              }
            }
          }
        },
        "summary": "Get all products",
        "tags": [
          "Products"
        ]
      },
      "post": {
        "operationId": "ProductController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateProductDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Product created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductResponseDto"
                }
              }
            }
          }
        },
        "summary": "Create a new product",
        "tags": [
          "Products"
        ]
      }
    },
    "/api/products/{id}": {
      "get": {
        "operationId": "ProductController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Product ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Product found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Product not found"
          }
        },
        "summary": "Get a product by ID",
        "tags": [
          "Products"
        ]
      },
      "patch": {
        "operationId": "ProductController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Product ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProductDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Product updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Product not found"
          }
        },
        "summary": "Update a product by ID",
        "tags": [
          "Products"
        ]
      },
      "delete": {
        "operationId": "ProductController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Product ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Product deleted"
          },
          "404": {
            "description": "Product not found"
          }
        },
        "summary": "Delete a product by ID",
        "tags": [
          "Products"
        ]
      }
    },
...
  },
  "info": {
    "title": "Stansbury Swim API",
    "description": "API documentation for Stansbury Swim backend",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "ScheduleStudentDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "birthDate": {
            "type": "string"
          },
          "notes": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "birthDate",
          "notes"
        ]
      },
      "ScheduleUserDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "firstName": {
            "type": "string"
          },
          "lastName": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "role": {
            "type": "string",
            "enum": [
              "user",
              "admin",
              "instructor"
            ]
          }
        },
        "required": [
          "id",
          "firstName",
          "lastName",
          "email",
          "role"
        ]
      },
      "ScheduleRegistrationDto": {
        "type": "object",
        "properties": {
          "studentId": {
            "type": "string"
          },
          "userId": {
            "type": "string"
          },
          "createdAt": {
            "type": "string"
          },
          "student": {
            "$ref": "#/components/schemas/ScheduleStudentDto"
          },
          "user": {
            "$ref": "#/components/schemas/ScheduleUserDto"
          }
        },
        "required": [
          "studentId",
          "userId",
          "createdAt",
          "student",
          "user"
        ]
      },
      "ScheduleResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "poolId": {
            "type": "string"
          },
          "instructorId": {
            "type": "string"
          },
          "classSize": {
            "type": "number"
          },
          "lessonType": {
            "type": "string",
            "enum": [
              "private",
              "group"
            ]
          },
          "startDateTime": {
            "type": "string"
          },
          "endDateTime": {
            "type": "string"
          },
          "registrations": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ScheduleRegistrationDto"
            }
          }
        },
        "required": [
          "id",
          "poolId",
          "instructorId",
          "classSize",
          "lessonType",
          "startDateTime",
          "endDateTime",
          "registrations"
        ]
      },
...
      "CreateScheduleDto": {
        "type": "object",
        "properties": {}
      },
...
      "CreateRegistrationDto": {
        "type": "object",
        "properties": {}
      },
      "PaymentResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          },
          "userId": {
            "type": "string"
          },
          "productId": {
            "type": "string"
          },
          "paymentGateway": {
            "type": "object"
          },
          "paymentGatewayId": {
            "type": "string"
          },
          "amount": {
            "type": "number"
          },
          "quantity": {
            "type": "number"
          },
          "status": {
            "type": "object"
          },
          "scheduleId": {
            "type": "string"
          },
          "studentId": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "createdAt",
          "updatedAt",
          "userId",
          "productId",
          "paymentGateway",
          "paymentGatewayId",
          "amount",
          "quantity",
          "status"
        ]
      },
...
      "CreatePaymentDto": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "string"
          },
          "productId": {
            "type": "string"
          },
          "quantity": {
            "type": "number"
          },
          "paymentGateway": {
            "type": "string",
            "enum": [
              "PAYPAL",
              "APPLE_PAY"
            ]
          },
          "paymentGatewayId": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "PENDING",
              "SUCCESS",
              "FAILED"
            ],
            "default": "PENDING"
          },
          "scheduleId": {
            "type": "string"
          },
          "studentId": {
            "type": "string"
          }
        },
        "required": [
          "userId",
          "productId",
          "quantity",
          "paymentGateway",
          "paymentGatewayId",
          "status"
        ]
      },
      "UpdatePaymentDto": {
        "type": "object",
        "properties": {}
      },
      "CreditBalanceDto": {
        "type": "object",
        "properties": {
          "creditType": {
            "type": "string",
            "enum": [
              "private",
              "group"
            ]
          },
          "balance": {
            "type": "number"
          }
        },
        "required": [
          "creditType",
          "balance"
        ]
      },
      "CreditBalanceResponseDto": {
        "type": "object",
        "properties": {
          "balances": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CreditBalanceDto"
            }
          }
        },
        "required": [
          "balances"
        ]
      },
      "TransactionResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "userId": {
            "type": "string"
          },
          "productId": {
            "type": "string"
          },
          "scheduleId": {
            "type": "string"
          },
          "transactionType": {
            "type": "string",
            "enum": [
              "PURCHASE_CREDITS",
              "REGISTER",
              "CANCEL_REGISTRATION"
            ]
          },
          "creditType": {
            "type": "string",
            "enum": [
              "private",
              "group"
            ]
          },
          "amount": {
            "type": "number"
          },
          "credits": {
            "type": "number"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          },
          "studentId": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "userId",
          "transactionType",
          "creditType",
          "credits",
          "createdAt",
          "updatedAt"
        ]
      },
      "CreateTransactionDto": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "string"
          },
          "productId": {
            "type": "string"
          },
          "scheduleId": {
            "type": "string"
          },
          "transactionType": {
            "type": "string",
            "enum": [
              "PURCHASE_CREDITS",
              "REGISTER",
              "CANCEL_REGISTRATION"
            ]
          },
          "creditType": {
            "type": "string",
            "enum": [
              "private",
              "group"
            ]
          },
          "credits": {
            "type": "number"
          },
          "amount": {
            "type": "number"
          },
          "paymentId": {
            "type": "string"
          },
          "studentId": {
            "type": "string"
          }
        },
        "required": [
          "userId",
          "transactionType",
          "creditType",
          "credits"
        ]
      },
      "ProductResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "productId123",
            "description": "Product ID"
          },
          "order": {
            "type": "number",
            "example": 1,
            "description": "Order of the product"
          },
          "name": {
            "type": "string",
            "example": "Product Name",
            "description": "Name of the product"
          },
          "lessonType": {
            "type": "string",
            "enum": [
              "private",
              "group"
            ],
            "description": "Type of lesson"
          },
          "credits": {
            "type": "number",
            "example": 10,
            "description": "Number of credits"
          },
          "active": {
            "type": "boolean",
            "example": true,
            "description": "Is the product active?"
          },
          "amount": {
            "type": "number",
            "example": 100,
            "description": "Amount for the product"
          },
          "description": {
            "type": "string",
            "example": "Description of the product",
            "description": "Product description"
          },
          "features": {
            "description": "List of product features",
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "id",
          "order",
          "name",
          "lessonType",
          "credits",
          "active",
          "amount",
          "description",
          "features"
        ]
      },
      "CreateProductDto": {
        "type": "object",
        "properties": {
          "order": {
            "type": "number",
            "example": 1,
            "description": "Order of the product"
          },
          "name": {
            "type": "string",
            "example": "Product Name",
            "description": "Name of the product"
          },
          "lessonType": {
            "type": "string",
            "enum": [
              "private",
              "group"
            ],
            "description": "Type of lesson"
          },
          "description": {
            "type": "string",
            "example": "Description of the product",
            "description": "Product description"
          },
          "credits": {
            "type": "number",
            "example": 10,
            "description": "Number of credits"
          },
          "amount": {
            "type": "number",
            "example": 100,
            "description": "Amount for the product"
          },
          "active": {
            "type": "boolean",
            "example": true,
            "description": "Is the product active?"
          },
          "scheduleId": {
            "type": "string",
            "example": "scheduleId123",
            "description": "Optional schedule ID"
          },
          "features": {
            "description": "List of product features",
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "order",
          "name",
          "lessonType",
          "description",
          "credits",
          "amount",
          "active",
          "features"
        ]
      },
      "UpdateProductDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "productId123",
            "description": "Product ID"
          },
          "name": {
            "type": "string",
            "example": "Product Name",
            "description": "Name of the product"
          },
          "lessonType": {
            "type": "string",
            "enum": [
              "private",
              "group"
            ],
            "description": "Type of lesson"
          },
          "credits": {
            "type": "number",
            "example": 10,
            "description": "Number of credits"
          },
          "active": {
            "type": "boolean",
            "example": true,
            "description": "Is the product active?"
          },
          "amount": {
            "type": "number",
            "example": 100,
            "description": "Amount for the product"
          },
          "description": {
            "type": "string",
            "example": "Description of the product",
            "description": "Product description"
          },
          "scheduleId": {
            "type": "string",
            "example": "scheduleId123",
            "description": "Optional schedule ID"
          }
        },
        "required": [
          "id"
        ]
      },
...
    }
  }
}